+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/prompt_guard/models/Message.java
@@ -1,0 +1,22 @@
+package cloud.pangeacyber.pangea.prompt_guard.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.NonNull;
+import lombok.Value;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@Value
+public final class Message {
+
+	@NonNull
+	String role;
+
+	@NonNull
+	String content;
+}