+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/prompt_guard/requests/GuardRequest.java
@@ -1,0 +1,26 @@
+package cloud.pangeacyber.pangea.prompt_guard.requests;
+
+import cloud.pangeacyber.pangea.BaseRequest;
+import cloud.pangeacyber.pangea.prompt_guard.models.Message;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.List;
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import lombok.Value;
+import lombok.experimental.SuperBuilder;
+import lombok.extern.jackson.Jacksonized;
+
+@EqualsAndHashCode(callSuper = true)
+@Jacksonized
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@SuperBuilder
+@Value
+public final class GuardRequest extends BaseRequest {
+
+	@NonNull
+	List<Message> messages;
+}