+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/prompt_guard/responses/GuardResponse.java
@@ -1,0 +1,6 @@
+package cloud.pangeacyber.pangea.prompt_guard.responses;
+
+import cloud.pangeacyber.pangea.Response;
+import cloud.pangeacyber.pangea.prompt_guard.results.GuardResult;
+
+public final class GuardResponse extends Response<GuardResult> {}